refactor: move utils to seperate files
