+++ lib/rules/vue-no-excessive-whitespace.js
@@ -1,60 +1,9 @@
-import { defineTemplateBodyVisitor, docsUrl } from '../utils/index.js'
+import { nodeParser, ruleMetaData, stringManipulation, templateProcessor } from '../utils/index.js'
 
-/**
- * Removes excessive whitespace from a class list string.
- *
- * @param {string} classList - The class list string to process.
- * @returns {string | void} - The class list string without excessive whitespace.
- */
-function removeWhitespace(classList) {
-  const classListWithWhitespace = classList.split(/(\s+)/)
-  const divider = ' '
-
-  const classListNoWhitespace = classListWithWhitespace.filter(
-    className => className.trim() !== '',
-  ).join(divider)
-
-  return classListNoWhitespace
-}
-
-function getNodeValue(node) {
-  if (node.type === 'TextAttribute' && node.name === 'class') {
-    return node.value
-  }
-  if (node.value === undefined) {
-    return node.value
-  }
-
-  switch (node.value.type) {
-    case 'JSXExpressionContainer':
-      return node.value.expression.value
-    case 'VExpressionContainer':
-      switch (node.value.expression.type) {
-        case 'ArrayExpression':
-          return node.value.expression.elements
-        case 'ObjectExpression':
-          return node.value.expression.properties
-      }
-      return node.value.expression.value
-    default:
-      return node.value.value
-  }
-}
-
-function getNodeRange(node) {
-  if (node.type === 'TextAttribute' && node.name === 'class') {
-    return [node.valueSpan.fullStart.offset, node.valueSpan.end.offset]
-  }
-  if (node.value === undefined) {
-    return [0, 0]
-  }
-  switch (node.value.type) {
-    case 'JSXExpressionContainer':
-      return node.value.expression.range
-    default:
-      return node.value.range
-  }
-}
+const { defineTemplateBodyVisitor } = templateProcessor()
+const { getDocsUrl } = ruleMetaData()
+const { getNodeValue, getNodeRange } = nodeParser()
+const { removeWhitespace } = stringManipulation()
 
 export default {
   meta: {
@@ -65,7 +14,7 @@ export default {
       description: 'Find excessive whitespace characters in class attribute in Vue templates.',
       recommended: true,
       category: 'Stylistic Issues',
-      url: docsUrl('vue-no-excessive-whitespace'),
+      url: getDocsUrl('vue-no-excessive-whitespace'),
     },
     messages: {
       'excessive-whitespace-in-class-attribute': 'The `class` attribute should not contain excessive whitespace.',
@@ -83,6 +32,7 @@ export default {
       if (argument === null) {
         originalClassNamesValue = getNodeValue(node)
         const range = getNodeRange(node)
+
         if (node.type === 'TextAttribute') {
           start = range[0]
           end = range[1]
@@ -110,7 +60,7 @@ export default {
 
               handleNodeArgument(node, propertyValue)
             })
-            return
+            break
           case 'Literal':
             originalClassNamesValue = argument.value
             start = argument.range[0] + 1
@@ -138,20 +88,23 @@ export default {
       }
     }
 
+    function callExpressionVisitor(node) {
+      const customCallees = ['clsx']
+
+      if (!customCallees.includes(node.callee.name))
+        return
+
+      node.arguments.forEach((argument) => {
+        handleNodeArgument(node, argument, 'excessive-whitespace-in-custom-callee')
+      })
+    }
+
     const scriptVisitor = {
-      CallExpression(node) {
-        const customCallees = ['clsx']
-
-        if (!customCallees.includes(node.callee.name))
-          return
-
-        node.arguments.forEach((argument) => {
-          handleNodeArgument(node, argument, 'excessive-whitespace-in-custom-callee')
-        })
-      },
+      CallExpression: callExpressionVisitor,
     }
 
     const templateVisitor = {
+      CallExpression: callExpressionVisitor,
       VAttribute(node) {
         switch (true) {
           case node.value && node.value.type === 'VLiteral':
